add arrow scroll js
