+++ src/assets/scripts/about.js
@@ -1,6 +1,8 @@
 import "./modules/skills";
 import fullMenu from "./modules/full-menu";
 import initMap from "./modules/map";
+import arrowScrollDown from "./modules/arrow-down";
 
 initMap();
 fullMenu();
+arrowScrollDown();
