+++ src/assets/scripts/blog.js
@@ -2,8 +2,10 @@ import fullMenu from "./modules/full-menu";
 import blogNav from "./modules/blog-nav";
 import blogArticle from "./modules/blog-article";
 import blogScroll from "./modules/blog-scroll";
+import arrowScrollDown from "./modules/arrow-down";
 
 blogNav();
 blogArticle();
 fullMenu();
 blogScroll();
+arrowScrollDown();
