+++ src/assets/scripts/works.js
@@ -1,4 +1,8 @@
 import fullMenu from "./modules/full-menu";
 import "./modules/slider";
+import arrowScrollDown from "./modules/arrow-down";
+import arrowScrollUp from "./modules/arrow-up";
 
 fullMenu();
+arrowScrollDown();
+arrowScrollUp();
