+++ src/views/pages/about.pug
@@ -25,9 +25,9 @@ block content
                         .author__info #{titleSite}
         .down-btn__hero
             .down-btn
-                a(href="#").down-btn__link
+                a(href="down").down-btn__link
                     +icon('portf_arrow_down', 'down-btn__icon')
-    section.about
+    section.about#down
         +triangles('about') 
         .about__left
             .about__left-title
