+++ src/views/pages/blog.pug
@@ -24,9 +24,9 @@ block content
                         .author__info #{titleSite}
         .down-btn__hero
             .down-btn
-                a(href="#").down-btn__link
+                a(href="down").down-btn__link
                     +icon('portf_arrow_down', 'down-btn__icon')
-    section.blog
+    section.blog#down
         +triangles('blog')
         .blog-column
             .blog-column__nav
